+++ apps/web/next.config.mjs
@@ -4,9 +4,9 @@ import { withWebSecurityHeaders } from '@pancakeswap/next-config/withWebSecurity
 import smartRouterPkgs from '@pancakeswap/smart-router/package.json' with { type: 'json' }
 import { withSentryConfig } from '@sentry/nextjs'
 import { createVanillaExtractPlugin } from '@vanilla-extract/next-plugin'
+import vercelToolbarPlugin from '@vercel/toolbar/plugins/next'
 import path from 'path'
 import { fileURLToPath } from 'url'
-import vercelToolbarPlugin from '@vercel/toolbar/plugins/next'
 
 const withVercelToolbar = vercelToolbarPlugin()
