+++ apps/web/src/state/achievements/helpers.ts
@@ -1,5 +1,5 @@
+import { campaignMap } from '@pancakeswap/achievements'
 import { TranslateFunction } from '@pancakeswap/localization'
-import { campaignMap } from 'config/constants/campaigns'
 import { GRAPH_API_PROFILE } from 'config/constants/endpoints'
 import { gql, request } from 'graphql-request'
 import { Achievement } from 'state/types'
