+++ apps/web/src/utils/achievements.ts
@@ -1,8 +1,8 @@
+import { campaignMap } from '@pancakeswap/achievements'
 import { ChainId } from '@pancakeswap/chains'
 import { getIfoConfig } from '@pancakeswap/ifos'
 import { TranslateFunction } from '@pancakeswap/localization'
 import { pointCenterIfoABI } from 'config/abi/pointCenterIfo'
-import { campaignMap } from 'config/constants/campaigns'
 import { Campaign, TranslatableText } from 'config/constants/types'
 import { Achievement } from 'state/types'
 import { getPointCenterIfoAddress } from 'utils/addressHelpers'
