+++ packages/achievements/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "@pancakeswap/tsconfig/base",
+  "include": ["./src"],
+  "exclude": ["**/*.test.ts"],
+  "compilerOptions": {
+    "target": "es2020",
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "skipLibCheck": true
+  }
+}
